# GridExpand/docs/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PreviewBody": {
        "additionalProperties": false,
        "description": "An edited copy of a base scenario: form ``changes`` (dotted keys) on top of ``text`` or the base file.",
        "properties": {
          "base": {
            "description": "file name from GET /api/scenarios",
            "maxLength": 200,
            "minLength": 1,
            "title": "Base",
            "type": "string"
          },
          "base_sha256": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "text_sha256 of GET …/form: refuse if the base file changed since",
            "title": "Base Sha256"
          },
          "changes": {
            "additionalProperties": true,
            "description": "dotted YAML path -> value (fields of GET …/form)",
            "maxProperties": 200,
            "title": "Changes",
            "type": "object"
          },
          "file_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "target file name (checked, not written)",
            "title": "File Name"
          },
          "scenario_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "sets scenario.id",
            "title": "Scenario Id"
          },
          "text": {
            "anyOf": [
              {
                "maxLength": 200000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "edited YAML text (default: the base file)",
            "title": "Text"
          }
        },
        "required": [
          "base"
        ],
        "title": "PreviewBody",
        "type": "object"
      },
      "SaveBody": {
        "additionalProperties": false,
        "description": "Save the edited copy as ``file_name`` in the user scenario directory.",
        "properties": {
          "base": {
            "description": "file name from GET /api/scenarios",
            "maxLength": 200,
            "minLength": 1,
            "title": "Base",
            "type": "string"
          },
          "base_sha256": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "text_sha256 of GET …/form: refuse if the base file changed since",
            "title": "Base Sha256"
          },
          "changes": {
            "additionalProperties": true,
            "description": "dotted YAML path -> value (fields of GET …/form)",
            "maxProperties": 200,
            "title": "Changes",
            "type": "object"
          },
          "file_name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "File Name",
            "type": "string"
          },
          "overwrite": {
            "default": false,
            "description": "replace an own file of the same name (a backup is kept)",
            "title": "Overwrite",
            "type": "boolean"
          },
          "scenario_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Scenario Id",
            "type": "string"
          },
          "text": {
            "anyOf": [
              {
                "maxLength": 200000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "edited YAML text (default: the base file)",
            "title": "Text"
          }
        },
        "required": [
          "base",
          "scenario_id",
          "file_name"
        ],
        "title": "SaveBody",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Jobs, results and scenario files of GridExpand (load allocation, urbs optimisation, power flow, grid expansion) for the GridPlanner UI. API version 1.",
    "title": "GridExpand API",
    "version": "api-1",
    "x-api-version": 1
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/scenarios": {
      "get": {
        "description": "Scenario YAMLs (``config/scenarios``, extra and user directories) with their key values.",
        "operationId": "list_scenarios_api_scenarios_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Scenarios Api Scenarios Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Scenarios",
        "tags": [
          "scenarios"
        ]
      },
      "post": {
        "description": "Save the edited copy as a new file in the user scenario directory.\n\nShipped files are never written; replacing an own file needs ``overwrite: true`` and keeps\n``<name>.bak-<timestamp>``. Returns the new ``GET /api/scenarios`` entry with ``backup``.",
        "operationId": "save_api_scenarios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Save Api Scenarios Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save",
        "tags": [
          "scenarios"
        ]
      }
    },
    "/api/scenarios/preview": {
      "post": {
        "description": "Apply the changes to a copy of the base file and validate it with the loader of the runs.\n\nNothing is written. Returns ``{ok, issues: [{level, message, key?, line?}], text, diff (unified),\nchanges, values, id, configuration_hash, scenario_key, unchanged, target}``.",
        "operationId": "preview_api_scenarios_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Preview Api Scenarios Preview Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview",
        "tags": [
          "scenarios"
        ]
      }
    },
    "/api/scenarios/{name}": {
      "delete": {
        "description": "Remove one of the user's scenario files (renamed to ``<name>.bak-<timestamp>``, not erased).",
        "operationId": "delete_api_scenarios__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Api Scenarios  Name  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete",
        "tags": [
          "scenarios"
        ]
      },
      "get": {
        "description": "The YAML text of one listed scenario file.",
        "operationId": "scenario_text_api_scenarios__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scenario Text",
        "tags": [
          "scenarios"
        ]
      }
    },
    "/api/scenarios/{name}/form": {
      "get": {
        "description": "Editable fields of a scenario file with their values and YAML help texts, and its text.\n\n``writable``: the file is one of the user's own (it can be replaced or deleted); ``user_dir``\ntells whether new files can be saved; ``proposal`` is the default file name and scenario id.",
        "operationId": "scenario_form_state_api_scenarios__name__form_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Scenario Form State Api Scenarios  Name  Form Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scenario Form State",
        "tags": [
          "scenarios"
        ]
      }
    },
...
  }
}